uia: move toggle/toggle_state to ButtonWrapper, move 'invoke' to UIAWrapper
